third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-script-reportonly-allowed-expected.txt

CONSOLE ERROR: line 5: The Content Security Policy 'require-sri-for script; script-src 'self' 'unsafe-inline'' was delivered via a <meta> element outside the document's <head>, which is disallowed. The policy has been ignored.
This is a testharness.js-based test.
PASS Script that requires integrity executes and does not generate a violation report. 
Harness: the test ran to completion.
